requirements.txt
aiohttp==3.6.2
aiotkinter==0.2
async-timeout==3.0.1
asyncio-periodic==2019.2
asyncqt==0.8.0
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
docopt==0.6.2
engineering-notation==0.6.0
friendly-traceback==0.0.39
future==0.18.2
idna==2.10
iniconfig==1.0.1
iso8601==0.1.12
kiwisolver==1.2.0
lxml==4.5.2
matplotlib==3.3.2
minimalmodbus==1.0.2
more-itertools==8.5.0
multidict==4.7.6
numpy==1.19.1
packaging==20.4
pandas==1.1.1
Pillow==7.2.0
pluggy==0.13.1
prompt-toolkit==2.0.4
protobuf==3.13.0
py==1.9.0
Pygments==2.2.0
pymodbus==2.3.0
pyparsing==2.4.7
PyQt5==5.13.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
pyqt5-tools==5.13.2.1.6rc1
PyQt5Designer==5.14.1
pyserial==3.4
PySide2==5.15.1
pytest==6.0.2
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pyuic5-tool==0.0.1
PyYAML==5.3.1
scipy==1.5.2
shiboken2==5.15.1
six==1.15.0
tk-tools==0.13.0
toml==0.10.1
tornado==6.0.4
wcwidth==0.2.5
xmltodict==0.12.0
yarl==1.5.1
